chore: move `service` result structs to dedicated file in `model` package
